feat: move domain files in entity package
