+++ src/main/java/com/turtleRun/be/running/domain/entity/RunningSessionStatistics.java
@@ -1,4 +1,4 @@
-package com.turtleRun.be.running.domain;
+package com.turtleRun.be.running.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -9,25 +9,28 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "running_splits")
+@Table(name = "running_session_statistics")
 @Getter
 @Setter
 @NoArgsConstructor
-public class RunningSplit {
+public class RunningSessionStatistics {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
+    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "session_id")
     private RunningSession session;
 
-    private Integer splitDistance;
-    private Integer splitDuration;
-    private BigDecimal splitPace;
+    private BigDecimal averageStrideLength;
+    private BigDecimal averageGroundContactTime;
+    private BigDecimal averageVerticalOscillation;
+    private BigDecimal averagePower;
+    private Integer averageCadence;
+    private Integer maxHeartRate;
     private Integer averageHeartRate;
-    private LocalDateTime startTime;
-    private LocalDateTime endTime;
+    private BigDecimal trainingEffect;
+    private BigDecimal vo2max;
     private LocalDateTime createdAt;
 
     @PrePersist
