+++ src/main/java/com/turtleRun/be/running/domain/entity/RunningSplit.java
@@ -1,4 +1,4 @@
-package com.turtleRun.be.running.domain;
+package com.turtleRun.be.running.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -9,28 +9,25 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "running_session_statistics")
+@Table(name = "running_splits")
 @Getter
 @Setter
 @NoArgsConstructor
-public class RunningSessionStatistics {
+public class RunningSplit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @OneToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "session_id")
     private RunningSession session;
 
-    private BigDecimal averageStrideLength;
-    private BigDecimal averageGroundContactTime;
-    private BigDecimal averageVerticalOscillation;
-    private BigDecimal averagePower;
-    private Integer averageCadence;
-    private Integer maxHeartRate;
+    private Integer splitDistance;
+    private Integer splitDuration;
+    private BigDecimal splitPace;
     private Integer averageHeartRate;
-    private BigDecimal trainingEffect;
-    private BigDecimal vo2max;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
     private LocalDateTime createdAt;
 
     @PrePersist
